plugins/geoclue2: Update plugin from the qt6 qtpositioning dev branch
Pull in the changes that were made to the plugin in the newest edition (qt6
dev branch of qtpositioning), and make a few adjustments so they still build
in our qt5 environment.

Note that the parseParameters() from the qt6 source is not used, because there
are no parameters parsed to the constructor in qt5. That means we keep the old
code to set the desktopId.