plugins/geoclue2: Use all available positioning methods by default
This plugin is broken when using it with aliendalvik/android appsupport. That's
because setPreferredPositioningMethods() isn't called by default, (the android
plugin is doing this, so is the winrt plugin, so was the old geoclue plugin),
so by default the positioning is very inaccurate.

So set the preferred positioning methods to "all" by default.