plugins/geoclue2: Remove guess of accuracy==0 means gps
Geoclue gives us a real accuracy, and doesn't just return 0 when GPS is
in use. This means that effectively m_lastPositionFromSatellite was likely
almost always "false", leading to lastKnownPosition() with
fromSatellitePositioningMethodsOnly returning nothing.

Instead let's do the same thing that the CoreLocation and the winrt plugin
are doing and ignore the fromSatellitePositioningMethodsOnly argument to
lastKnownPosition().